test-libs import issues fix.
Now func include is done in pytonic way.
If __init__.func is present in the directory
it is included, otherwise fallback to the old way.
Drawback is that if __init__.func is present,
one HAS TO add "#include" line there, otherwise
new file won't be included.

Maybe we should do both?
Include __init__ and then include everything
in the directory anyways?
Crude, but should to the trick.